requirements.txt
Flask
paho-mqtt
sqlalchemy
psycopg2-binary
